+++ EmployeeManagement.Application/CQRS/Positions/Handlers/Commands/DeleteUserCommandHandler.cs
@@ -7,6 +7,8 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using EmployeeManagement.Application.CQRS.Positions.Requests.Commands;
+using EmployeeManagement.Application.Exceptions;
+using EmployeeManagement.Domain;
 
 namespace EmployeeManagement.Application.CQRS.Positions.Handlers.Commands
 {
@@ -24,6 +26,11 @@ namespace EmployeeManagement.Application.CQRS.Positions.Handlers.Commands
             public async Task<Unit> Handle(DeletePositionCommand request, CancellationToken cancellationToken)
             {
                 var position = await _positionRepository.GetAsync(request.Id);
+
+                if (position == null)
+                {
+                    throw new NotFoundException(nameof(Position), request.Id);
+                }
 
                 await _positionRepository.DeleteAsync(position);
